dht2: Modify glusterd to create DHT2 based volumes
This is a throwaway modification, which still holds some glusterd
concepts in tact, to be able to have some modicum of usability.

This adds the ability to create volumes using DHT2 as follows,
gluster volume create <volname> mds <count> data <count> ...

Here, mds is the count of meta data volumes (better name needed),
and data is the count of data volumes that DHT2 needs to use. The
brick list that follows should total (after considering how many
subvolumes to DHT2 is formed, i.e considering replica/ec/etc.)
the sum of mds and data counts specified.

Just specifying these options during create would shift the volume
to a DHT2 based volume, replacing DHT in the volume graph as needed.

The changes also prevent further add/remove bricks as the current
prototype does not intend to handle the same.

Other operations like rebalance are left as is, and should not be
used.

This change is to set the stage further to just prototype DHT2 and
test it out with better ease than hand wiring vol files or
mucking around with DHT code.

Change-Id: I613d88265e4058ee92c91abe625183e3cd9a6406
Signed-off-by: ShyamsundarR <srangana@redhat.com>