dht2: Initial conceptual design for the second DHT2 prototype
The initial DHT2 design keeps the file inode and data blocks in
the DS subvolumes. This makes lookup, readdirp, creat, do 2 RPCs
before being able to complete the request.

The second prototype as presented here, separats the data blocks
from the file inode, and attempts to retain the file inode in the
same subvolume (even on rebalance) as its first parent.

This optimizes the calls as mentioned above, but also brings about
requirements to enable keeping the size/times updated on the inode
when data changes in the DS.

This thought is being currently enhanced to see how viable it is
and if the benefits are tangible.

Change-Id: I0e5bbf9157281a3e1305ac1d1e0685af1409b7da
Signed-off-by: ShyamsundarR <srangana@redhat.com>