cluster/dht2: Prototype implementation of lookup and stat FOPs
This is just a prototype implementation of lookup and stat to
exchange notes on DHT2 and POSIX2 development streams and to
come to a consensus on how to proceed on certain areas.

The intention further would be to use the code generation
techniquies like NSR to generate swathes of DHT2 FOPs, both
to speed up the development and also the accrue the benefits
of the code generation method of writing xlators.

Change-Id: I504bab66fdb1b6177919809faee03ebf671a9daf
Signed-off-by: Shyam <srangana@redhat.com>