cluster/dht2: Prototype implementation of lookup FOP
This is just a prototype implementation of lookup to
exchange notes on DHT2 and POSIX2 development streams and to
come to a consensus on how to proceed on certain areas.

These sets of prototype commits, lead to an actual POC for DHT2,
at which point, these would be moved into *real* production
code.

Change-Id: I504bab66fdb1b6177919809faee03ebf671a9daf
Signed-off-by: Shyam <srangana@redhat.com>