Update patch set 2
Patch Set 2:

Updated code to remove EREMOTE reason as an xattr, rather just treat EREMOTE as a remote inode operation and move on. So that this complies with the POSIX2 implementation and that is what we need.

Also, modified loc when doing just inode lookups in the EREMOTE case, so that we do not confuse the lower xlators, and also POSIX can work cleaner when it knows we are just operating on the inode and not path etc.

Patch-set: 2
Label: Code-Review=0