Update patch set 3
Patch Set 3:

> Shyam,
 > 
 > Do you plan to incorporate layout handling changes too (in this
 > patch or thereafter, considering that it would be useful in the
 > future)? Let me know, I can take care of those changes.

Yes, I will incorporate this in the (near?) future (short answer).

Currently I am not blocked on this and also the points at which the layoutxxx functions are used are minimal, so replacing these with a *common* interface is easier at present.

Also, I would like some more areas where we use the layout functions and then generate an abstraction for the same.

Some doubts on the abstraction remain though, for example,
- dht2_find_metdata_server_subvol understands we are working with buckets derived out of gfid, this is an abstraction by itself (and anyway someone should know ;) )
- More importantly, what are the inargs for layout search? Say it is conf and loc, then we can abstract how we work with a loc, extracting the GFID->bucket etc. This way if a different layout is sought, then these inargs are used. But I need to see some more examples for the level at which we need this abstraction, before adding it to the code.

Say Tier the inargs may consider FOP type, if create redirect to *hot* in which case inargs become different.

In short (again), I think this can wait and we can solidify this better and then incorporate it.

Patch-set: 3
Label: Code-Review=0