cluster/dht2: Prototype implementation of create
This is a prototype implementation of create FOP in DHT2.

It compiles, and needs some testing with POSIX2 implementation of
the same FOPs. Additionally it is limited in implementation, it
will not handle EREMOTE create cases as detailed in the comments

These sets of prototype commits, lead to an actual POC for DHT2,
at which point, these would be moved into *real* production code.

Change-Id: Iec4c52f0cae28810df82d12b3d9443d8ab6effdc
Signed-off-by: Shyam <srangana@redhat.com>