cluster/dht2: Add codegen based prototype open/stat/setattr/flush FOPs
Core code for the mentioned FOPs are added. This set is required
for create to work from a FUSE mount as FUSE sends in these FOPs
on a create.

The FOPs are added using code gen methods introduced in code gluster
for default routines. This can hence be exended to other common
inode and fd based FOPs that need to be written subsequently.

These sets of prototype commits, lead to an actual POC for DHT2,
at which point, these would be moved into *real* production code.

Change-Id: If6874c9aaa269d1f16e1f707edade5d116007c70
Signed-off-by: Shyam <srangana@redhat.com>