cluster/dht2: Add codegen based prototype FOPs
Added a series of common FOPs, inode and fd based FOPs like stat,
truncate, set/get/remove(xattr), etc. Also added some FOPs that
would not be supported in the POC implementation.

The FOPs are added using code gen methods introduced in code gluster
for default routines. This can hence be exended to other common
inode and fd based FOPs that need to be written subsequently.

This is part of a set of prototype commits that can lead to and actual
POC for DHT2

Change-Id: If6874c9aaa269d1f16e1f707edade5d116007c70
Signed-off-by: Shyam <srangana@redhat.com>