Integration: Fixes for remote inode lookup to work
Some integration fixes that would be absorbed into individual
commits, for remote inode lookups to work.

Primarily, on the client and server protocol ends, an error
of EREMOTE needed handling and passing back of recieved iatt
information to xlators above.

Something similar to the same in the new POSIX store was needed
to pass back iatt on EREMOTE errors.

Change-Id: I54bc5b444c0e8a0f66ee4064d1c7274ad3911ce4
Signed-off-by: Shyam <srangana@redhat.com>