Update patch set 1
Patch Set 1:

Very much needed change.

Minor nit: Would it make sense to have an helper routine for remote object checks given that now some of of the translators need to know about this special errno? That way it's easier for any part of the xlator stack to not have similar checks or have the knowledge of EREMOTE but just take an action depending on whether the looked up object is remote or not. This should be a simplistic change (can be done later, but better late than never :))

int gf_is_object_remote (int op_errno)
{
        return (op_errno == EREMOTE);
}

Patch-set: 1