posix2/ds: Data Server as separate xlator
This patches contains skeleton code to make separate
posix xlator called "posix2-ds" for Data Server stack.
It also involves glusterd changes to load "posix2-mds"
if it is a MDS brick and "posix2-ds" if it's a DS brick
on to a server stack.

In DS, on lookup, do icreate.

Change-Id: I6fe4906818bc5aee57340776aaf0dd82adf0f15a
Signed-off-by: Kotresh HR <kotreshhr5@gmail.com>